requirements.txt
streamlit
groq
pymupdf
langchain
langchain-text-splitters
faiss-cpu
sentence-transformers
torch
torchvision
python-dotenv
